requirements.txt
streamlit>=1.28
pandas>=2.0
numpy>=1.24
scikit-learn==1.6.1
joblib>=1.3
matplotlib>=3.8
seaborn>=0.13
python-dotenv>=1.0
google-generativeai>=0.3.0
requests>=2.31
plotly>=5.17
